+++ contracts/btc-staking/src/msg.rs
@@ -105,4 +105,6 @@ pub struct FinalityProviderInfo {
     pub total_active_sats: u64,
     /// Whether this finality provider is slashed
     pub slashed: bool,
+    /// Height that other info is indexed by
+    pub height: u64,
 }